@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.17.1 <3.0.0"
 
 dependencies:
   flutter:
@@ -30,26 +30,26 @@ dependencies:
   cupertino_icons: ^1.0.0
   google_sign_in: ^5.0.4
   flutter_icons: ^1.1.0
-  percent_indicator: ^3.0.1
+  percent_indicator: ^4.2.2
   cached_network_image: ^3.0.0
   flutter_spinkit: ^5.0.0
-  cloud_firestore: ^2.3.0
+  cloud_firestore: ^3.4.3
   fluttertoast: ^8.0.7
   animations: ^2.0.0
   image: ^3.0.2
   uuid: ^3.0.4
-  firebase_auth: ^2.0.0
+  firebase_auth: ^3.6.2
   firebase_core: ^1.3.0
   image_picker: ^0.8.1+1
-  firebase_storage: ^9.0.0
+  firebase_storage: ^10.3.4
   path_provider: ^2.0.2
-  geolocator: ^7.2.0+1
+  geolocator: ^9.0.1
   geocoding: ^2.0.0
   timeago: ^3.1.0
-  google_fonts: ^2.1.0
-  provider: ^5.0.0
+  google_fonts: ^3.0.1
+  provider: ^6.0.3
   modal_progress_hud: ^0.1.3
-  image_cropper: ^1.3.1
+  image_cropper: ^3.0.0
   shared_preferences: ^2.0.6
   data_connection_checker: ^0.3.4
 
