@@ -4,18 +4,18 @@ import 'package:flutter/material.dart';
 import '../posts/create_post.dart';
 
 class FabContainer extends StatelessWidget {
-  final Widget page;
+  final Widget? page;
   final IconData icon;
   final bool mini;
 
-  FabContainer({@required this.page, @required this.icon, this.mini = false});
+  FabContainer({ this.page, required this.icon, this.mini = false});
 
   @override
   Widget build(BuildContext context) {
     return OpenContainer(
       transitionType: ContainerTransitionType.fade,
       openBuilder: (BuildContext context, VoidCallback _) {
-        return page;
+        return page!;
       },
       closedElevation: 4.0,
       closedShape: const RoundedRectangleBorder(
