@@ -3,20 +3,20 @@ import 'package:flutter_icons/flutter_icons.dart';
 import 'package:social_media_app/components/custom_card.dart';
 
 class PasswordFormBuilder extends StatefulWidget {
-  final String initialValue;
-  final bool enabled;
-  final String hintText;
-  final TextInputType textInputType;
-  final TextEditingController controller;
-  final TextInputAction textInputAction;
-  final FocusNode focusNode, nextFocusNode;
-  final VoidCallback submitAction;
+  final String? initialValue;
+  final bool? enabled;
+  final String? hintText;
+  final TextInputType? textInputType;
+  final TextEditingController? controller;
+  final TextInputAction? textInputAction;
+  final FocusNode? focusNode, nextFocusNode;
+  final VoidCallback? submitAction;
   final bool obscureText;
-  final FormFieldValidator<String> validateFunction;
-  final void Function(String) onSaved, onChange;
-  final Key key;
-  final IconData prefix;
-  final IconData suffix;
+  final FormFieldValidator<String>? validateFunction;
+  final void Function(String)? onSaved, onChange;
+  final Key? key;
+  final IconData? prefix;
+  final IconData? suffix;
 
   PasswordFormBuilder(
       {this.prefix,
@@ -41,7 +41,7 @@ class PasswordFormBuilder extends StatefulWidget {
 }
 
 class _PasswordFormBuilderState extends State<PasswordFormBuilder> {
-  String error;
+  String? error;
   bool obscureText = false;
 
   @override
@@ -56,16 +56,17 @@ class _PasswordFormBuilderState extends State<PasswordFormBuilder> {
             child: Container(
               child: Theme(
                 data: ThemeData(
-                  primaryColor: Theme.of(context).accentColor,
-                  accentColor: Theme.of(context).accentColor,
+                  primaryColor: Theme.of(context).colorScheme.secondary,
+                  colorScheme: ColorScheme.fromSwatch().copyWith(
+                      secondary: Theme.of(context).colorScheme.secondary),
                 ),
                 child: TextFormField(
                   initialValue: widget.initialValue,
                   enabled: widget.enabled,
                   onChanged: (val) {
-                    error = widget.validateFunction(val);
+                    error = widget.validateFunction!(val);
                     setState(() {});
-                    widget.onSaved(val);
+                    widget.onSaved!(val);
                   },
                   style: TextStyle(
                     fontSize: 15.0,
@@ -77,18 +78,18 @@ class _PasswordFormBuilderState extends State<PasswordFormBuilder> {
                   keyboardType: widget.textInputType,
                   validator: widget.validateFunction,
                   onSaved: (val) {
-                    error = widget.validateFunction(val);
+                    error = widget.validateFunction!(val);
                     setState(() {});
-                    widget.onSaved(val);
+                    widget.onSaved!(val!);
                   },
                   textInputAction: widget.textInputAction,
                   focusNode: widget.focusNode,
                   onFieldSubmitted: (String term) {
                     if (widget.nextFocusNode != null) {
-                      widget.focusNode.unfocus();
+                      widget.focusNode!.unfocus();
                       FocusScope.of(context).requestFocus(widget.nextFocusNode);
                     } else {
-                      widget.submitAction();
+                      widget.submitAction!();
                     }
                   },
                   decoration: InputDecoration(
