@@ -1,14 +1,14 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 
 class PostModel {
-  String id;
-  String postId;
-  String ownerId;
-  String username;
-  String location;
-  String description;
-  String mediaUrl;
-  Timestamp timestamp;
+  String? id;
+  String? postId;
+  String? ownerId;
+  String? username;
+  String? location;
+  String? description;
+  String? mediaUrl;
+  Timestamp? timestamp;
   
 
   PostModel({
