@@ -1,11 +1,11 @@
 
 class Register {
-  String username;
-  String email;
-  String gender;
-  String country;
-  String password;
-  String passwordConfirmation;
+  String? username;
+  String? email;
+  String? gender;
+  String? country;
+  String? password;
+  String? passwordConfirmation;
   bool publicEmail = false;
 
   Map<String, dynamic> toJson() {
