@@ -12,7 +12,7 @@ class Activities extends StatefulWidget {
 
 class _ActivitiesState extends State<Activities> {
   currentUserId() {
-    return firebaseAuth.currentUser.uid;
+    return firebaseAuth.currentUser!.uid;
   }
 
   @override
@@ -32,7 +32,7 @@ class _ActivitiesState extends State<Activities> {
                 style: TextStyle(
                   fontSize: 13.0,
                   fontWeight: FontWeight.w900,
-                  color: Theme.of(context).accentColor,
+                  color: Theme.of(context).colorScheme.secondary,
                 ),
               ),
             ),
@@ -59,7 +59,8 @@ class _ActivitiesState extends State<Activities> {
           .snapshots(),
       physics: NeverScrollableScrollPhysics(),
       itemBuilder: (_, DocumentSnapshot snapshot) {
-        ActivityModel activities = ActivityModel.fromJson(snapshot.data());
+        ActivityModel activities =
+            ActivityModel.fromJson(snapshot.data() as Map<String, dynamic>);
         return ActivityItems(
           activity: activities,
         );
@@ -70,7 +71,7 @@ class _ActivitiesState extends State<Activities> {
   deleteAllItems() async {
 //delete all notifications associated with the authenticated user
     QuerySnapshot notificationsSnap = await notificationRef
-        .doc(firebaseAuth.currentUser.uid)
+        .doc(firebaseAuth.currentUser!.uid)
         .collection('notifications')
         .get();
     notificationsSnap.docs.forEach(
