@@ -11,7 +11,7 @@ import 'package:social_media_app/widgets/cached_image.dart';
 import 'package:timeago/timeago.dart' as timeago;
 
 class Comments extends StatefulWidget {
-  final PostModel post;
+  final PostModel? post;
 
   Comments({this.post});
 
@@ -19,14 +19,14 @@ class Comments extends StatefulWidget {
 }
 
 class _CommentsState extends State<Comments> {
-  UserModel user;
+  UserModel? user;
 
   PostService services = PostService();
   final DateTime timestamp = DateTime.now();
   TextEditingController commentsTEC = TextEditingController();
 
   currentUserId() {
-    return firebaseAuth.currentUser.uid;
+    return firebaseAuth.currentUser!.uid;
   }
 
   @override
@@ -82,7 +82,7 @@ class _CommentsState extends State<Comments> {
                           controller: commentsTEC,
                           style: TextStyle(
                             fontSize: 15.0,
-                            color: Theme.of(context).textTheme.headline6.color,
+                            color: Theme.of(context).textTheme.headline6!.color,
                           ),
                           decoration: InputDecoration(
                             contentPadding: EdgeInsets.all(10.0),
@@ -102,7 +102,7 @@ class _CommentsState extends State<Comments> {
                             hintStyle: TextStyle(
                               fontSize: 15.0,
                               color:
-                                  Theme.of(context).textTheme.headline6.color,
+                                  Theme.of(context).textTheme.headline6!.color,
                             ),
                           ),
                           maxLines: null,
@@ -112,9 +112,9 @@ class _CommentsState extends State<Comments> {
                             await services.uploadComment(
                               currentUserId(),
                               commentsTEC.text,
-                              widget.post.postId,
-                              widget.post.ownerId,
-                              widget.post.mediaUrl,
+                              widget.post!.postId!,
+                              widget.post!.ownerId!,
+                              widget.post!.mediaUrl!,
                             );
                             commentsTEC.clear();
                           },
@@ -122,7 +122,7 @@ class _CommentsState extends State<Comments> {
                             padding: const EdgeInsets.only(right: 10.0),
                             child: Icon(
                               Icons.send,
-                              color: Theme.of(context).accentColor,
+                              color: Theme.of(context).colorScheme.secondary,
                             ),
                           ),
                         ),
@@ -147,7 +147,7 @@ class _CommentsState extends State<Comments> {
         Container(
           height: 250.0,
           width: MediaQuery.of(context).size.width - 20.0,
-          child: cachedNetworkImage(widget.post.mediaUrl),
+          child: cachedNetworkImage(widget.post!.mediaUrl!),
         ),
         Padding(
           padding: const EdgeInsets.all(8.0),
@@ -158,7 +158,7 @@ class _CommentsState extends State<Comments> {
                 crossAxisAlignment: CrossAxisAlignment.start,
                 children: [
                   Text(
-                    widget.post.description,
+                    widget.post!.description!,
                     style: TextStyle(
                       fontWeight: FontWeight.w800,
                     ),
@@ -167,20 +167,20 @@ class _CommentsState extends State<Comments> {
                   Row(
                     children: [
                       Text(
-                        timeago.format(widget.post.timestamp.toDate()),
+                        timeago.format(widget.post!.timestamp!.toDate()),
                         style: TextStyle(),
                       ),
                       SizedBox(width: 3.0),
                       StreamBuilder(
                         stream: likesRef
-                            .where('postId', isEqualTo: widget.post.postId)
+                            .where('postId', isEqualTo: widget.post!.postId)
                             .snapshots(),
                         builder:
                             (context, AsyncSnapshot<QuerySnapshot> snapshot) {
                           if (snapshot.hasData) {
-                            QuerySnapshot snap = snapshot.data;
+                            QuerySnapshot snap = snapshot.data!;
                             List<DocumentSnapshot> docs = snap.docs;
-                            return buildLikesCount(context, docs?.length ?? 0);
+                            return buildLikesCount(context, docs.length ?? 0);
                           } else {
                             return buildLikesCount(context, 0);
                           }
@@ -204,13 +204,14 @@ class _CommentsState extends State<Comments> {
       shrinkWrap: true,
       padding: const EdgeInsets.symmetric(horizontal: 20.0),
       stream: commentRef
-          .doc(widget.post.postId)
+          .doc(widget.post!.postId)
           .collection('comments')
           .orderBy('timestamp', descending: true)
           .snapshots(),
       physics: NeverScrollableScrollPhysics(),
       itemBuilder: (_, DocumentSnapshot snapshot) {
-        CommentModel comments = CommentModel.fromJson(snapshot.data());
+        CommentModel comments =
+            CommentModel.fromJson(snapshot.data() as Map<String, dynamic>);
         return Column(
           crossAxisAlignment: CrossAxisAlignment.start,
           mainAxisAlignment: MainAxisAlignment.start,
@@ -219,21 +220,21 @@ class _CommentsState extends State<Comments> {
               contentPadding: EdgeInsets.symmetric(horizontal: 20.0),
               leading: CircleAvatar(
                 radius: 20.0,
-                backgroundImage: NetworkImage(comments.userDp),
+                backgroundImage: NetworkImage(comments.userDp!),
               ),
               title: Text(
-                comments.username,
+                comments.username!,
                 style: TextStyle(fontWeight: FontWeight.w700),
               ),
               subtitle: Text(
-                timeago.format(comments.timestamp.toDate()),
+                timeago.format(comments.timestamp!.toDate()),
                 style: TextStyle(fontSize: 12.0),
               ),
             ),
             Padding(
               padding: const EdgeInsets.symmetric(horizontal: 20.0),
               child: Text(
-                comments.comment,
+                comments.comment!,
                 style: TextStyle(fontWeight: FontWeight.w400),
               ),
             ),
@@ -247,18 +248,18 @@ class _CommentsState extends State<Comments> {
   buildLikeButton() {
     return StreamBuilder(
       stream: likesRef
-          .where('postId', isEqualTo: widget.post.postId)
+          .where('postId', isEqualTo: widget.post!.postId)
           .where('userId', isEqualTo: currentUserId())
           .snapshots(),
       builder: (context, AsyncSnapshot<QuerySnapshot> snapshot) {
         if (snapshot.hasData) {
-          List<QueryDocumentSnapshot> docs = snapshot?.data?.docs ?? [];
+          List<QueryDocumentSnapshot> docs = snapshot.data?.docs ?? [];
           return IconButton(
             onPressed: () {
               if (docs.isEmpty) {
                 likesRef.add({
                   'userId': currentUserId(),
-                  'postId': widget.post.postId,
+                  'postId': widget.post!.postId,
                   'dateCreated': Timestamp.now(),
                 });
                 addLikesToNotification();
@@ -297,37 +298,37 @@ class _CommentsState extends State<Comments> {
   }
 
   addLikesToNotification() async {
-    bool isNotMe = currentUserId() != widget.post.ownerId;
+    bool isNotMe = currentUserId() != widget.post!.ownerId;
 
     if (isNotMe) {
       DocumentSnapshot doc = await usersRef.doc(currentUserId()).get();
-      user = UserModel.fromJson(doc.data());
+      user = UserModel.fromJson(doc.data() as Map<String, dynamic>);
       notificationRef
-          .doc(widget.post.ownerId)
+          .doc(widget.post!.ownerId)
           .collection('notifications')
-          .doc(widget.post.postId)
+          .doc(widget.post!.postId)
           .set({
         "type": "like",
-        "username": user.username,
+        "username": user!.username!,
         "userId": currentUserId(),
-        "userDp": user.photoUrl,
-        "postId": widget.post.postId,
-        "mediaUrl": widget.post.mediaUrl,
+        "userDp": user!.photoUrl!,
+        "postId": widget.post!.postId,
+        "mediaUrl": widget.post!.mediaUrl,
         "timestamp": timestamp,
       });
     }
   }
 
   removeLikeFromNotification() async {
-    bool isNotMe = currentUserId() != widget.post.ownerId;
+    bool isNotMe = currentUserId() != widget.post!.ownerId;
 
     if (isNotMe) {
       DocumentSnapshot doc = await usersRef.doc(currentUserId()).get();
-      user = UserModel.fromJson(doc.data());
+      user = UserModel.fromJson(doc.data() as Map<String, dynamic>);
       notificationRef
-          .doc(widget.post.ownerId)
+          .doc(widget.post!.ownerId)
           .collection('notifications')
-          .doc(widget.post.postId)
+          .doc(widget.post!.postId)
           .get()
           .then((doc) => {
                 if (doc.exists) {doc.reference.delete()}
