@@ -11,7 +11,7 @@ import 'package:social_media_app/view_models/profile/edit_profile_view_model.dar
 import 'package:social_media_app/widgets/indicators.dart';
 
 class EditProfile extends StatefulWidget {
-  final UserModel user;
+  final UserModel? user;
 
   const EditProfile({this.user});
 
@@ -20,10 +20,10 @@ class EditProfile extends StatefulWidget {
 }
 
 class _EditProfileState extends State<EditProfile> {
-  UserModel user;
+  UserModel? user;
 
   String currentUid() {
-    return firebaseAuth.currentUser.uid;
+    return firebaseAuth.currentUser!.uid;
   }
 
   @override
@@ -48,7 +48,7 @@ class _EditProfileState extends State<EditProfile> {
                     style: TextStyle(
                       fontWeight: FontWeight.w900,
                       fontSize: 15.0,
-                      color: Theme.of(context).accentColor,
+                      color: Theme.of(context).colorScheme.secondary,
                     ),
                   ),
                 ),
@@ -82,7 +82,7 @@ class _EditProfileState extends State<EditProfile> {
                           padding: const EdgeInsets.all(1.0),
                           child: CircleAvatar(
                             radius: 65.0,
-                            backgroundImage: NetworkImage(viewModel.imgLink),
+                            backgroundImage: NetworkImage(viewModel.imgLink!),
                           ),
                         )
                       : viewModel.image == null
@@ -91,14 +91,14 @@ class _EditProfileState extends State<EditProfile> {
                               child: CircleAvatar(
                                 radius: 65.0,
                                 backgroundImage:
-                                    NetworkImage(widget.user.photoUrl),
+                                    NetworkImage(widget.user!.photoUrl!),
                               ),
                             )
                           : Padding(
                               padding: const EdgeInsets.all(1.0),
                               child: CircleAvatar(
                                 radius: 65.0,
-                                backgroundImage: FileImage(viewModel.image),
+                                backgroundImage: FileImage(viewModel.image!),
                               ),
                             ),
                 ),
@@ -124,7 +124,7 @@ class _EditProfileState extends State<EditProfile> {
           children: [
             TextFormBuilder(
               enabled: !viewModel.loading,
-              initialValue: widget.user.username,
+              initialValue: widget.user!.username,
               prefix: Feather.user,
               hintText: "Username",
               textInputAction: TextInputAction.next,
@@ -135,7 +135,7 @@ class _EditProfileState extends State<EditProfile> {
             ),
             SizedBox(height: 10.0),
             TextFormBuilder(
-              initialValue: widget.user.country,
+              initialValue: widget.user!.country,
               enabled: !viewModel.loading,
               prefix: Feather.map_pin,
               hintText: "Country",
@@ -152,16 +152,16 @@ class _EditProfileState extends State<EditProfile> {
             ),
             TextFormField(
               maxLines: null,
-              initialValue: widget.user.bio,
+              initialValue: widget.user!.bio,
               autovalidateMode: AutovalidateMode.onUserInteraction,
-              validator: (String value) {
-                if (value.length > 1000) {
+              validator: (String? value) {
+                if (value!.length > 1000) {
                   return 'Bio must be short';
                 }
                 return null;
               },
-              onSaved: (String val) {
-                viewModel.setBio(val);
+              onSaved: (String? val) {
+                viewModel.setBio(val!);
               },
               onChanged: (String val) {
                 viewModel.setBio(val);
