@@ -45,7 +45,7 @@ class _TabScreenState extends State<TabScreen> {
     {
       'title': 'Profile',
       'icon': CupertinoIcons.person,
-      'page': Profile(profileId: firebaseAuth.currentUser.uid),
+      'page': Profile(profileId: firebaseAuth.currentUser!.uid),
       'index': 4,
     },
   ];
@@ -84,7 +84,7 @@ class _TabScreenState extends State<TabScreen> {
                           item['icon'],
                           color: item['index'] != _page
                               ? Colors.grey
-                              : Theme.of(context).accentColor,
+                              : Theme.of(context).colorScheme.secondary,
                           size: 20.0,
                         ),
                         onPressed: () => navigationTapped(item['index']),
