@@ -1,7 +1,7 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:social_media_app/utils/constants.dart';
+import 'package:social_media_app/view_models/theme/theme_view_model.dart';
 
 class Setting extends StatefulWidget {
   @override
@@ -31,16 +31,17 @@ class _SettingState extends State<Setting> {
         padding: EdgeInsets.all(10.0),
         child: ListView(
           children: <Widget>[
-             ListTile(
-              title: Text(
-                "About",
-                style: TextStyle(
-                  fontWeight: FontWeight.w900,
+            ListTile(
+                title: Text(
+                  "About",
+                  style: TextStyle(
+                    fontWeight: FontWeight.w900,
+                  ),
                 ),
-              ),
-              subtitle: Text("A Fully Functional Social Media Application Made by CharlyKeleb"),
-              trailing: Icon(Icons.error)
-            ),
+                subtitle: Text(
+                  "A Fully Functional Social Media Application Made by CharlyKeleb",
+                ),
+                trailing: Icon(Icons.error)),
             Divider(),
             ListTile(
               title: Text(
@@ -50,13 +51,13 @@ class _SettingState extends State<Setting> {
                 ),
               ),
               subtitle: Text("Use the dark mode"),
-              trailing: Consumer<ThemeNotifier>(
+              trailing: Consumer<ThemeProvider>(
                 builder: (context, notifier, child) => CupertinoSwitch(
                   onChanged: (val) {
                     notifier.toggleTheme();
                   },
                   value: notifier.dark,
-                  activeColor: Theme.of(context).accentColor,
+                  activeColor: Theme.of(context).colorScheme.secondary,
                 ),
               ),
             ),
