@@ -2,7 +2,7 @@ import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/foundation.dart';
 
 class UserViewModel extends ChangeNotifier {
-  User user;
+  User? user;
   FirebaseAuth auth = FirebaseAuth.instance;
 
   setUser() {
