@@ -5,7 +5,7 @@ import 'package:social_media_app/screens/view_image.dart';
 import 'package:social_media_app/widgets/cached_image.dart';
 
 class PostTile extends StatefulWidget {
-  final PostModel post;
+  final PostModel? post;
 
   PostTile({this.post});
 
@@ -34,7 +34,7 @@ class _PostTileState extends State<PostTile> {
             borderRadius: BorderRadius.all(
               Radius.circular(3.0),
             ),
-            child: cachedNetworkImage(widget.post.mediaUrl),
+            child: cachedNetworkImage(widget.post!.mediaUrl!),
           ),
         ),
       ),
