@@ -23,37 +23,37 @@ class Chats extends StatelessWidget {
         ),
         title: Text("Chats"),
       ),
-      body: StreamBuilder(
-        stream: userChatsStream('${viewModel.user?.uid ?? ""}'),
+      body: StreamBuilder<QuerySnapshot>(
+        stream: userChatsStream('${viewModel.user!.uid ?? ""}'),
         builder: (context, snapshot) {
           if (snapshot.hasData) {
-            List chatList = snapshot.data.docs;
+            List chatList = snapshot.data!.docs;
             if (chatList.isNotEmpty) {
               return ListView.separated(
                 itemCount: chatList.length,
                 itemBuilder: (BuildContext context, int index) {
                   DocumentSnapshot chatListSnapshot = chatList[index];
-                  return StreamBuilder(
+                  return StreamBuilder<QuerySnapshot>(
                     stream: messageListStream(chatListSnapshot.id),
                     builder: (context, snapshot) {
                       if (snapshot.hasData) {
-                        List messages = snapshot.data.docs;
+                        List messages = snapshot.data!.docs;
                         Message message = Message.fromJson(
                           messages.first.data(),
                         );
                         List users = chatListSnapshot.get('users');
                         // remove the current user's id from the Users
                         // list so we can get the second user's id
-                        users.remove('${viewModel.user?.uid ?? ""}');
+                        users.remove('${viewModel.user!.uid ?? ""}');
                         String recipient = users[0];
                         return ChatItem(
                           userId: recipient,
-                          messageCount: messages?.length,
-                          msg: message?.content,
-                          time: message?.time,
+                          messageCount: messages.length,
+                          msg: message.content!,
+                          time: message.time!,
                           chatId: chatListSnapshot.id,
-                          type: message?.type,
-                          currentUserId: viewModel.user?.uid ?? "",
+                          type: message.type!,
+                          currentUserId: viewModel.user!.uid ?? "",
                         );
                       } else {
                         return SizedBox();
