@@ -5,6 +5,7 @@ import 'package:social_media_app/view_models/auth/posts_view_model.dart';
 import 'package:social_media_app/view_models/auth/register_view_model.dart';
 import 'package:social_media_app/view_models/conversation/conversation_view_model.dart';
 import 'package:social_media_app/view_models/profile/edit_profile_view_model.dart';
+import 'package:social_media_app/view_models/theme/theme_view_model.dart';
 import 'package:social_media_app/view_models/user/user_view_model.dart';
 import 'package:social_media_app/utils/constants.dart';
 
@@ -15,5 +16,5 @@ List<SingleChildWidget> providers = [
   ChangeNotifierProvider(create: (_) => EditProfileViewModel()),
   ChangeNotifierProvider(create: (_) => ConversationViewModel()),
   ChangeNotifierProvider(create: (_) => UserViewModel()),
-  ChangeNotifierProvider(create: (_) => ThemeNotifier()),
+  ChangeNotifierProvider(create: (_) => ThemeProvider()),
 ];
