@@ -10,13 +10,13 @@ class LoginViewModel extends ChangeNotifier {
   GlobalKey<FormState> formKey = GlobalKey<FormState>();
   bool validate = false;
   bool loading = false;
-  String email, password;
+  String? email, password;
   FocusNode emailFN = FocusNode();
   FocusNode passFN = FocusNode();
   AuthService auth = AuthService();
 
   login(BuildContext context) async {
-    FormState form = formKey.currentState;
+    FormState form = formKey.currentState!;
     form.save();
     if (!form.validate()) {
       validate = true;
@@ -49,14 +49,14 @@ class LoginViewModel extends ChangeNotifier {
   forgotPassword(BuildContext context) async {
     loading = true;
     notifyListeners();
-    FormState form = formKey.currentState;
+    FormState form = formKey.currentState!;
     form.save();
     print(Validations.validateEmail(email));
     if (Validations.validateEmail(email) != null) {
       showInSnackBar('Please input a valid email to reset your password.',context);
     } else {
       try {
-        await auth.forgotPassword(email);
+        await auth.forgotPassword(email!);
         showInSnackBar('Please check your email for instructions '
             'to reset your password', context);
       } catch (e) {
